refactor(test): move session handler tests to a sibling file
The inline test module had grown past the ~500-line threshold at which
tests move to foo/tests.rs. Test bodies are unchanged; the module is
still a child of handlers::session, so private access is unaffected.
